refactor: Alterar nome do dir. nav-bar para NavBar
