+++ src/layouts/GlobalLayout.jsx
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom'
-import { NavBar } from '../components/nav-bar'
+import { NavBar } from '../components/NavBar'
 
 export default function GlobalLayout() {
     return (
